+++ threedigrid_builder/base/__init__.py
@@ -1,5 +1,6 @@
-from .array import array_of  # NOQA
+from .array import *  # NOQA
 from .lines import *  # NOQA
 from .nodes import *  # NOQA
 from .pumps import *  # NOQA
 from .interfaces import *  # NOQA
+from .settings import *  # NOQA